requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
python-dotenv>=0.19.1
openai>=1.0.0
google-genai>=1.0.0
gunicorn>=20.1.0
